Expand scope of QtGamepad to QtUniversalInput
This patch transforms the QtGamepad module from Qt 5 into a new
module that expands the scope to include all forms of exotic input.

The majority of the previous QtGamepad functionality and APIs are
preserved, but now there is also a more general API for getting
the state of input devices like joysticks, flight yokes, and
any other devices that define themselves in terms of buttons and
axes.

In addition to being able to receive input, it is also now possible
to output feedback in devices that support it.

Experimental input mapping functionality has been added to facilitate
mapping an arbitrary input to a predefined action.

Change-Id: Ic35b73282944cfd4eea9defbcf82c4abc55d7237
Reviewed-by: Andy Nichols <andy.nichols@qt.io>